+++ doc/versionHistory.rst
@@ -6,6 +6,15 @@
 Version History
 ##################
 
+.. _lsst.ts.wep-14.1.2:
+
+-------------
+14.1.2
+-------------
+
+* Fix test data by adjusting ROTPA values in headers after DM-49838.
+* Enforce additional possible single threading options.
+
 .. _lsst.ts.wep-14.1.1:
 
 -------------